+++ PlantUmlViewer/PlantUmlViewer/Forms/OptionsWindow.cs
@@ -16,11 +16,13 @@ namespace PlantUmlViewer.Forms
             InitializeComponent();
 
             textBox_JavaPath.Text = settings.Settings.JavaPath;
+            numericUpDown_ExportSizeFactor.Value = settings.Settings.ExportSizeFactor;
         }
 
         private void Button_Ok_Click(object sender, EventArgs e)
         {
             settings.Settings.JavaPath = textBox_JavaPath.Text;
+            settings.Settings.ExportSizeFactor = numericUpDown_ExportSizeFactor.Value;
             settings.Save();
 
             DialogResult = DialogResult.OK;