+++ PlantUmlViewer/PlantUmlViewer/Settings/PlantUmlViewerSettings.cs
@@ -3,5 +3,6 @@
     public class PlantUmlViewerSettings
     {
         public string JavaPath { get; set; } = "";
+        public decimal ExportSizeFactor { get; set; } = 1;
     }
 }