+++ tools/create_release.py
@@ -35,7 +35,7 @@ def create_plugin_list_entry(version, architecture, hash):
         'id': '{hash}'.format(hash=hash),
         'repository': 'https://github.com/Fruchtzwerg94/PlantUmlViewer/releases/download/{version}/PlantUmlViewer_v{version}_{architecture}.zip'
             .format(version=version, architecture=architecture),
-        'description': 'A Notepad++ plugin to generate view and export PlantUML diagrams.',
+        'description': 'A Notepad++ plugin to generate, view and export PlantUML diagrams.',
         'author': 'Philipp Schmidt',
         'homepage': 'https://github.com/Fruchtzwerg94/PlantUmlViewer'
     }